imsdk/crop-resize-gpu/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.3.0
PyGObject==3.50.0
